clang-format: improve Default style for crate and format all files
Improve the Default style for the clang-format crate so that if
a .clang-format file exists that doesn't override the "Default".
Also add a File style so that if users want to use the .clang-format
style they use this.

Add a .clang-format file that defaults our project to Mozilla,
and add a .clang-format which disables formatting for doctest.

Then format all .cpp and .h files with clang-format using Mozilla.

Change-Id: I7068259ccb8a074691000ffcc84c15d9281972ab
Reviewed-on: https://codereview.kdab.com/c/kdab/rust/cxx-qt/+/98450
Reviewed-by: Gerhard de Clercq <gerhard.declercq@kdab.com>