cxx-qt-lib: implement UniquePtrTarget for opaque types
This allows us later to pass around cxx::UniquePtr<T>
where T is the opaque type later.

Which then allows us to return T so that we don't need
the pimpl implementation of the opaque types
(eg QVariant, Variant, VariantImpl).

Which was required as Variant was returned to C++ so needed
to be trivial.

Once we have UniquePtr<T> returning this returns QVariant
so on the Rust side all we have is Variant.

Change-Id: I29a8bec3ee869a794fd013a146e1ac8f30ad1145