qcolor: use inner unique_ptr instead of stack object
This simplifies the cxx-qt-lib qcolor as we don't need to handle the creation of a stack object via macros and instead initialise the object on the C++ and keep a handle to a unique_ptr of it on the Rust side. Then when we return across the bridge we pass the unique_ptr across. Any interaction with the C++ object then uses methods which take the unique_ptr.