cxx-qt-build: decouple code generation from file writing
also cleanup ambiguous usage of str/String with Path/PathBuf

The locations of generated files have been changed for consistency
and simplifying the code. Previously, C++ code generated by cxx was
put in files named based on the Rust file name and the C++ header
file was confusingly put in the src directory. This has been changed
so all files use the snake_cased version of the cxx module name,
with the files generated by cxx having a _cxx suffix. Both cxx's
and cxx-qt's generated headers are now in the include directory.