build Qt types autogenerated C++ code as a library
Previously, the C++ code was autogenerated by cxx-qt-build from
applications' build.rs, then repeatedly recompiled for each
application. Now, the C++ code is generated once and built as
a library which is linked to each application. This is a
prerequisite for removing the need for applications to have
build.rs to instead use the CLI C++ code generator. Also, this
avoids redundant compilation of the autogenerated C++ which is
constant acrosss applications.