generate C++ with CLI tool instead of build.rs
This moves the build of the Rust libraries from CMake configure time to
CMake build time, resolving https://github.com/KDAB/cxx-qt/issues/106.

In addition to behaving in a more conventional manner for a C++ build
system, this reduces the build time of this repository by avoiding
redundant compilation of the Rust code that does the C++ code
generation.