cxx-qt-gen: add a parser module for creating an abstract QObject
This module will replace the extract.rs, taking syn ItemMod as an
input and then rearranging the Item's into an abstract
representation of the QObject's in the module.

Later a generate module will then take this abstract representation
and convert it into a RustObject and a CppObject, which will then
be passed to a writer module to create source files.