remove cxx-qt-cli crate
This was a good try for improving the build process:
https://github.com/KDAB/cxx-qt/pull/173
However, the practicalities of distributing the Rust exectuable,
plus a C++ library, plus a CMake module, would not be worth it.
Now that cxx-qt-build has been refactored so it compiles the
generated C++ code from build.rs scripts, there is no need for a
separate CLI tool.