cxx-qt-gen: use impl qobject::T to define the C++ context
This then creates a natural way to add methods to the C++ context.

It also means later with nested objects they can be referred to
by crates::module::qobject::MyObject.

And any Rust traits or methods are natural as they are impl A for T,
with traits for the C++ object being natural with impl A for qobject::T