cxx-qt-gen: use the module name or attribute for the CXX file names
So that we can support multiple QObjects later in one module block,
use the module name or cxx_stem attribute as the output file name.

Later if the Rust API is ever stable we could use the inspection
APIs on Span to read the filename of the macro. This would mean
we would be the same as CXX but this requires nightly.