cxx-qt-gen: add wrapper struct for closure to be queue()-ed
The idea is to pass a (callback fn, Send + 'static closure) pair in to
QMetaObject::invokeMethod(), and invoke the callback fn with that closure
as an argument. The callback fn just invokes the passed closure. This way,
we can queue Rust closure back to QObject thread.

    thread::spawn(move || {
        ...
        let response = String::new();
        qt_thread.queue(move |qobject| qobject.emit(Finished { response }));
    });

This should be simpler and less error-prone than sending result back via
a separate channel.

This commit adds an opaque struct for the Send + 'static closure.