Use new RFC 2795 formating from Rust 1.58
This allows us to avoid the dreaded passing of trailing args to format
strings and instead let's us embed them directly in the string itself.

Note that it is not (yet) possible to embed arbitrary expressions in the
format string, for example for the following we still have to use the normal syntax:
println!("cargo:rustc-link-search={}", dir.display());

Otherwise Rust will complain with:
invalid format string: expected `'}'`, found `'.'`if you intended to print `{`, you can escape it using `{{`

However, at least for normal variables this works and makes up for quite
the nice cleanup.

Fixes #24