Implement Eq for all basic Qt types
This is an addendum to 540670065afc084a7ffbf532941a502e5538182c

It enables these types to be used as a type in hashmaps.

Implementing Eq only requires that the type's equality operator is an
equivalence relation, which in this case dumps down to being PartialEq
AND reflexive.

The reflexive property does hold for most types, but not for all, for
example we can't implement it for any type that compares floats as the
IEEE requires that NaN is not equal to NaN, which breaks the reflexive
property.