cxx-qt: don't use unwrap and convert to a compile error instead
This improve error messages when our cxx-qt-gen code breaks.

Eg I have added a duplicate attribute key-value in the code,
before we used to get this error when compiling.

```
error: custom attribute panicked
  --> examples/cargo_without_cmake/src/cxxqt_object.rs:11:1
   |
11 | #[cxx_qt::bridge]
   | ^^^^^^^^^^^^^^^^^
   |
   = help: message: called `Result::unwrap()` on an `Err` value: Error("Duplicate keys in the attributes")
```

Now we have a much better error that points to the code block that is at fault.

```
error: Duplicate keys in the attributes
  --> examples/cargo_without_cmake/src/cxxqt_object.rs:23:5
   |
23 |     #[cxx_qt::qobject(qml_uri = "com.kdab.cxx_qt.demo", qml_version = "1.0", qml_version = "1.0")]
   |     ^
```

Related to #38