cxx-qt-build: do not use -bundle otherwise lib doesn't end up in CMake
When building a staticlib -bundle means that the native static library
is simply not included into the archive and some higher level build
system will need to add it later during linking of the final binary.
https://doc.rust-lang.org/rustc/command-line-arguments.html#option-l-link-lib

Otherwise when we build with CMake the libqt-static-initializers.a
is not bundled into the final libqt_minimal.a so is then missing
from the final binary.

Closes #597