cxx-qt-build: add Cargo feature to toggle linking Qt's .o files
Unfortunately, there is no way to get linking to work in every case
without conditional compilation. This feature needs to be enabled
for bin crates or when running `cargo test` for library crates. The
feature must not be enabled when builing a staticlib then linking
it with CMake because CMake will also link the .o files, which
makes linking fail with duplicate symbol errors.