Extract former CXXMappings into TypeNames struct
This is the first stage of the Naming phase (#835) refactor.
All associations between types and their fully qualified names are now
stored in one central database, the TypeNames struct.

The final goal for this struct is to include all fully qualified names
for each nameable type in the bridge.
Therefore the generator can use this to lookup any references from e.g.
arguments.

This is not quite done yet, as this isn't yet used everywhere, but it's
a good first step