CI: Always update the compiler cache after merging
Previously, it would skip updating, as the key is always the same.
That meant the cache got progressively older.

Now every commit that lands will update the cache.
All forks will use the same cache from main, which keeps the cache alive.
Because the cache should update often, it should be close enough for
most forks to still be cached effectively.