fix(ci): Use whole-archive only for initializers
Previously, we linked the entire staticlib generated by the Rust
compiler into our CMake target.
This was part of what caused the recent issues in CI (see #958).
Cargo already only linked the plugin&resources library as whole-archive,
which is where where it is needed.
Now we simply export the initializers library into the CXXQT_EXPORT_DIR,
so we can only link the initializers with whole-archive.
This is a lot cleaner and should lead to fewer linker failures.

Closes #958