Remove cxx_qt::Locking
This has turned out to cause the most difficult issues with #1034.

cxx_qt::Locking has always been a bit of an unnecessary precaution.
A normal Qt integration should not break the Rust safety guarantees for
the following reasons:
- Qt objects are always assigned to a single thread
  - They should always be destructed from this thread
  - They should always be modified form this thread
- We support this pattern with CxxQtThread, which queues onto the same
  thread

So protection from multi-threaded access isn't really necessary, as all
access and modification should be single-threaded.

Removing Locking allows us to close #1034, and should improve
performance, as locking a mutex for every call isn't all that cheap.
It also simplifies generation, as we don't have to generate wrappers
anymore.

If C++ manages to const_cast or some other nasty behavior, that's really
bad C++ code, which we can't protect from anyway.
As per our current policy, we consider any C++ code to be unsafe. As
long as you cannot create unsafe behavior from *safe* code, that is good
enough. So this fits with the policy.