CXX-Qt-build: Remove definitions from Interface
These definitions cannot be imported into CMake, as they are unknown at
configure time.
It is also best practice to keep definitions in a header file, as that
allows us to only include them where really needed, as pointed out by
this blog post: https://www.kdab.com/setting-defines-with-cmake/

This is the method that cxx-qt-lib now uses to configure its features.

Closes #1165