create individual qvector header files
The files were created because the Rust bridges, which are specific, included qvector, which is generic.
This, in turn, included all possible types of qvector, even though only the type required by the bridge was actually needed.
(Run-time: n^2)
With the change, each spefific bridge now has its own specific qvector that it can include.

The only downside is, that the user has to manually include the specific qvector header.