meta_project -> qml_multi_crates & enable CMake
qml_multi_crates is the grand finale in our quest for CMake&Cargo
compatibility.
It is a project that can be built as a Cargo lib and binary, as well as
imported into CMake with a C++ binary and uses multiple Cargo crates
internally.

It does rely on the init_crate! and init_qml_module! macros, but
otherwise works quite seamless.