Multi-Crate project (#1178)
* examples: Multiple crate project

* meta_project: specify and document Rust 1.74 MSRV

This isn't required for the libraries published to crates.io,
so this isn't specified in the workspace Cargo.toml.

* meta_project: use `extern crate` rather than `pub use`

No need to change the privacy of these; just need to have the
symbols referenced within the staticlib crate.

* Allow multi-crate projects with new build system

This also adds cxx_qt::init_crate! and init_qml_module!

* Update book with Common Issues section

* Improve CMake compatibility with multiple crates

Previously, downstream QML modules weren't exported.
This should now be fixed, as QML modules are always exported.

* meta_project -> qml_multi_crates & enable CMake

qml_multi_crates is the grand finale in our quest for CMake&Cargo
compatibility.
It is a project that can be built as a Cargo lib and binary, as well as
imported into CMake with a C++ binary and uses multiple Cargo crates
internally.

It does rely on the init_crate! and init_qml_module! macros, but
otherwise works quite seamless.

* Update Changelog and documentation

* Initialize crates QML modules in crate initializer

---------

Co-authored-by: Andrew Hayzen <andrew.hayzen@kdab.com>
Co-authored-by: Be Wilson <be.0@gmx.com>