cxx-qt-lib: create individual header files
* create individual header files
The files were created because the Rust bridges, which are specific, included headers, which is generic.
This, in turn, included all possible types, even though only the type required by the bridge was actually needed.
(Run-time n^2)
With the change, each spefific bridge now has its own specific header that it can include.

The only downside is, that the user has to manually include the specific header.

* Add missing header files

* Add missing gitattributes
- set qset/qlist-headers to autogenerated/linguist-generated

* remove code duplication

* fixing clang-format errors

* fixing clang-format errors

---------

Co-authored-by: Quentin Weber <quentin.weber@kdab.com>