cxx-qt-lib: Add two more QtLogging functions
This adds two new global functions available in <QtLogging>:
qFormatLogMessage and qSetMessagePattern.

qFormatLogMessage is similar to qt_message_output, but actually
documented. Instead of going to the default message handler though, it
outputs to a QString. This is probably not useful without message
handler support, but it's easy to expose.

qSetMessagePattern is very useful, as it allows CXX-Qt applications to
set custom Qt message patterns for logging.