CXX-Qt-build: Simplify QmlModule (#1342)
* Remove QRC files from QML modules

This is for the most part already covered by the .qrc function in
CxxQtBuilder.

* CXX-Qt-Build: Move QmlModule to builder pattern

* CXX-Qt-Build: Pass QmlModule into new

As we now only allow a single QML module per builder, we can just pass
it directly into the new function.

* CXX-Qt-build: Add CxxQtBuilder::qrc_resources

This is our version of qt_add_resources and replaces the previous
qrc_files function in QmlModule

* build: Use QmlUri everywhere

* Remove duplicate qrc file from demo_threading

* Warn about duplicate Rust files in builder

This seems to have caused the linker errors we experienced in CI.
Also fix the demo_threading build script, which had this issue.

* cargo_without_cmake: Remove useless qrc_files call