Object type declarations in qmldir-files (#1349)
* added BlueRect.qml to the qml_multi_crates example.

Added the file to the sub1 subproject and a call to it in main.qml.

* now the qmldir generator output object type declarations

* Use local paths for qml files in QmlDirBuilder

The QmlDirBuilder should only build the contents of the qmldir files,
and not modify any files on disk.
Ensuring the files are in the correct (relative) paths of the QML module
is the responsibility of register_qml_module.

* Add license header to BlueRect.qml

* Do not use OsStr::display due to MSRV

* Remove version form QmlDirBuilder

These should be specified per-file, and not for the entire module

---------

Co-authored-by: Leon Matthes <leon.matthes@kdab.com>